% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareClustersByTrt.R
\name{CompareClustersByTrt}
\alias{CompareClustersByTrt}
\title{Given a Seurat object and treatment variable, compare
clusters between levels of the treatment}
\usage{
CompareClustersByTrt(
  obj,
  trt_var,
  rep_var = NULL,
  group_by = NULL,
  ci = TRUE,
  ci_alpha = 0.05,
  seed = 1
)
}
\arguments{
\item{obj}{Seurat object}

\item{trt_var}{Treatment variable}

\item{rep_var}{Replicate/Individual ID variable}

\item{group_by}{Name of grouping variable (default is Idents)}

\item{ci}{Whether to plot binomial confidence intervals (Jeffreys)}
}
\description{
Given a Seurat object and treatment variable, compare
clusters between levels of the treatment
}
\examples{
pbmc_small$trt <- sample(c('drug', 'control'), ncol(pbmc_small), replace=TRUE)
pbmc_small$genotype <- sample(c('1', '2', '3'), ncol(pbmc_small), replace=TRUE)
CompareClustersByTrt(pbmc_small, quo(trt), quo(genotype)) + xlab('')
}
