% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteCellex.R
\name{WriteCELLEX}
\alias{WriteCELLEX}
\title{Write data in CELLEX format https://github.com/perslab/CELLEX.}
\usage{
WriteCELLEX(seurat_obj, cellranger_dir, outprefix, outdir = ".")
}
\arguments{
\item{seurat_obj}{Seurat object to print}

\item{cellranger_dir}{Dir of Cellranger results. Needed to get ENS IDs}

\item{outprefix}{Character. Prefix for outfiles}

\item{outdir}{Character. Directory in which to place files}
}
\description{
Write data in CELLEX format https://github.com/perslab/CELLEX.
}
\examples{
WriteCELLEX(seurat_obj, cellranger_dir, outprefix, outdir)
}
