% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write10X.R
\name{Write10X}
\alias{Write10X}
\title{Write data in 10X format i.e. directory with barcodes.tsv, genes.tsv, matrix.mtx}
\usage{
Write10X(obj, dir)
}
\description{
Write data in 10X format i.e. directory with barcodes.tsv, genes.tsv, matrix.mtx
}
\examples{
Write10X(obj, dir)
}
