% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write10X.R
\name{Write10X}
\alias{Write10X}
\title{Write data in 10X format i.e. directory with barcodes.tsv, genes.tsv, matrix.mtx
If object is Seurat v2, output cellranger2-like format
If object is Seurat v3, output cellranger3-like format}
\usage{
Write10X(obj, dir)
}
\arguments{
\item{obj}{Seurat object to print}

\item{dir}{Character. Directory in which to place 10X-like files}
}
\description{
Write data in 10X format i.e. directory with barcodes.tsv, genes.tsv, matrix.mtx
If object is Seurat v2, output cellranger2-like format
If object is Seurat v3, output cellranger3-like format
}
\examples{
Write10X(obj, dir)
}
